Acc emp moved to area folders
