+++ C#.NET Apps/YouTubeProjects/YTP.Main/Areas/HumanResources/Models/AccEmpData.cs
@@ -1,18 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace YTP.Main.ViewModel {
+namespace YTP.Main.Areas.HumanResources.Models {
     public class AccEmpData {
 
         public int EmployeeId { get; set; }
 
-        [Display(Name = "First Name: "), Required(ErrorMessage ="First Name field is required. ")]
+        [Display(Name = "First Name: "), Required(ErrorMessage = "First Name field is required. ")]
         public string FirstName { get; set; }
 
-        [Display(Name = "Last Name: "), Required(ErrorMessage ="Last Name field is required. ")]
+        [Display(Name = "Last Name: "), Required(ErrorMessage = "Last Name field is required. ")]
         public string LastName { get; set; }
 
         [Display(Name = "Department: "), Required(ErrorMessage = "Department field is required. ")]
