+++ C#.NET Apps/YouTubeProjects/YTP.Main/Areas/HumanResources/Views/Shared/_Layout.cshtml
@@ -23,7 +23,7 @@
                         @Html.ActionLink("Service #1", "Index", "HaitiEmployee", null, new { @class = "nav-link" })
                     </li>
                     <li class="nav-item">
-                        @Html.ActionLink("Service #2", "Index", "Services", null, new { @class = "nav-link" })
+                        @Html.ActionLink("Service #2", "Index", "AccEmp", null, new { @class = "nav-link" })
                     </li>
                     <li class="nav-item">
                         @Html.ActionLink("Service #3", "Index", "Services", null, new { @class = "nav-link" })
@@ -42,6 +42,7 @@
                             <li class="nav-item"><a class="nav-link disabled">All Projects</a></li>
                             <li><hr class="dropdown-divider"></li>
                             <li>@Html.ActionLink("Haiti Employee", "Index", "HaitiEmployee", new { area = "HumanResources" }, new { @class = "nav-link dropdown-item" })</li>
+                            <li>@Html.ActionLink("Acc Employee", "Index", "AccEmp", new { area = "HumanResources" }, new { @class = "nav-link dropdown-item" })</li>
                         </ul>
                     </li>
                     <li class="nav-item">
