+++ C#.NET Apps/YouTubeProjects/YTP.Main/Views/Shared/_Layout.cshtml
@@ -46,7 +46,7 @@
                             <li>@Html.ActionLink("Northwind Employees", "Index", "NorthwindEmployee", new { area = "" }, new { @class = "nav-link dropdown-item" })</li>
                             <li>@Html.ActionLink("Product Management", "Index", "Product", new { area = "" }, new { @class = "nav-link dropdown-item" })</li>
                             <li>@Html.ActionLink("TBS Employee", "Index", "TBS_Employee", new { area = "" }, new { @class = "nav-link" })</li>
-                            <li>@Html.ActionLink("Acc Employee", "Index", "AccEmp", new { area = "" }, new { @class = "nav-link" })</li>
+                            <li>@Html.ActionLink("Acc Employee", "Index", "AccEmp", new { area = "HumanResources" }, new { @class = "nav-link" })</li>
                         </ul>
                     </li>
                     <li class="nav-item"><a class="nav-link disabled">Admin</a></li>
