+++ C#.NET Apps/YouTubeProjects/YTP.Main/YTP.Main.csproj
@@ -143,7 +143,7 @@
     <Compile Include="Controllers\EmployeeController.cs" />
     <Compile Include="Controllers\ErrorController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
-    <Compile Include="Controllers\NorthwindEmployeeController.cs" />
+    <Compile Include="Areas\HumanResources\Controllers\NorthwindEmployeeController.cs" />
     <Compile Include="Controllers\ProductController.cs" />
     <Compile Include="Controllers\TBS_EmployeeController.cs" />
     <Compile Include="DataAccess\Category.cs">
@@ -187,7 +187,7 @@
     </Compile>
     <Compile Include="DataAccess\Sp_EmpDBAccess.cs" />
     <Compile Include="Models\Employee.cs" />
-    <Compile Include="Models\NorthwindEmployees.cs" />
+    <Compile Include="Areas\HumanResources\Models\NorthwindEmployees.cs" />
     <Compile Include="DataAccess\Product.cs">
       <DependentUpon>DefaultDBAccess.tt</DependentUpon>
     </Compile>
@@ -332,7 +332,7 @@
     <Content Include="Views\Employee\Create.cshtml" />
     <Content Include="Views\Employee\Index.cshtml" />
     <Content Include="Views\Employee\Edit.cshtml" />
-    <Content Include="Views\NorthwindEmployee\Index.cshtml" />
+    <Content Include="Areas\HumanResources\Views\NorthwindEmployee\Index.cshtml" />
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\Product\Index.cshtml" />
     <Content Include="Views\TBS_Employee\Index.cshtml" />
